src/main/java/com/christianj98/pomodoro/service/implementation/LoginServiceImpl.java
package com.christianj98.pomodoro.service.implementation;

import com.christianj98.pomodoro.dto.AuthRequestDto;
import com.christianj98.pomodoro.dto.JwtResponseDto;
import com.christianj98.pomodoro.service.JwtService;
import com.christianj98.pomodoro.service.LoginService;
import lombok.RequiredArgsConstructor;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class LoginServiceImpl implements LoginService {

    private final AuthenticationManager authenticationManager;
    private final JwtService jwtService;

    @Override
    public Authentication authenticate(AuthRequestDto authRequestDto) {
        return authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authRequestDto.getUsername(), authRequestDto.getPassword()));
    }

    @Override
    public JwtResponseDto generateTokenForAuthenticatedUser(Authentication authentication, String username) {
        if (authentication.isAuthenticated()) {
            return JwtResponseDto.builder()
                    .accessToken(jwtService.generateToken(username))
                    .build();
        } else {
            throw new UsernameNotFoundException("Invalid user request..!!");
        }
    }
}
